Issue #77. Refactoring. JsonDictionary. Move JSON parsing to the lexical unit classes.
